Fixing validation for number and country code
When an invalid number or country code were entered, the VATValidator would incorrectly return true. I assume the return true on VATExceptions is to allow inaccessible VIES service scenarios, however this also allows invalid codes such as: GB1 or X#1

This fix allows the invalid numbers and codes to be caught and an error shown, before even checking the VIES service.